Move side buttons widget to widget package
