+++ src/main/java/slimeknights/tconstruct/tables/client/inventory/widget/SideButtonsWidget.java
@@ -1,4 +1,4 @@
-package slimeknights.tconstruct.tables.client.inventory.module;
+package slimeknights.tconstruct.tables.client.inventory.widget;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.gui.components.Button;
